+++ internal/services/core/domain/file/file_service_test.go
@@ -1,4 +1,4 @@
-package service
+package corefile
 
 import (
 	"bytes"